+++ test/cases/array-only.js
@@ -1,9 +1,9 @@
 var types = require('../types');
-var position = types.position;
+var location = types.location;
 var array = types.createArray;
 
 module.exports = {
-	ast: array([], position(1, 1, 0, 1, 3, 2)),
+	ast: array([], location(1, 1, 0, 1, 3, 2)),
 	options: {
 		verbose: true
 	}