+++ test/cases/deep.js
@@ -1,25 +1,25 @@
 var types = require('../types');
 var key = types.createObjectKey;
-var position = types.position;
+var location = types.location;
 var object = types.createObject;
 var prop = types.createObjectProperty;
 var array = types.createArray;
 var string = types.createString;
 
-var n = array([string('n')]/*, position()*/);
-var m = array([string('m'), n]/*, position()*/);
-var l = array([string('l'), m]/*, position()*/);
-var k = array([string('k'), l]/*, position()*/);
-var j = array([string('j'), k]/*, position()*/);
-var i = array([string('i'), j]/*, position()*/);
-var h = array([string('h'), i]/*, position()*/);
-var g = object([prop(key('g'), h)]/*, position()*/);
-var f = object([prop(key('f'), g)]/*, position()*/);
-var e = object([prop(key('e'), f)]/*, position()*/);
-var d = object([prop(key('d'), e)]/*, position()*/);
-var c = object([prop(key('c'), d)]/*, position()*/);
-var b = object([prop(key('b'/*, position(3, 5, 15, 3, 8, 18)*/), c)]/*, position()*/);
-var a = object([prop(key('a'/*, position(2, 3, 4, 2, 6, 7)*/), b)]/*, position(1, 1, 0, 29, 2, 516)*/);
+var n = array([string('n')]/*, location()*/);
+var m = array([string('m'), n]/*, location()*/);
+var l = array([string('l'), m]/*, location()*/);
+var k = array([string('k'), l]/*, location()*/);
+var j = array([string('j'), k]/*, location()*/);
+var i = array([string('i'), j]/*, location()*/);
+var h = array([string('h'), i]/*, location()*/);
+var g = object([prop(key('g'), h)]/*, location()*/);
+var f = object([prop(key('f'), g)]/*, location()*/);
+var e = object([prop(key('e'), f)]/*, location()*/);
+var d = object([prop(key('d'), e)]/*, location()*/);
+var c = object([prop(key('c'), d)]/*, location()*/);
+var b = object([prop(key('b'/*, location(3, 5, 15, 3, 8, 18)*/), c)]/*, location()*/);
+var a = object([prop(key('a'/*, location(2, 3, 4, 2, 6, 7)*/), b)]/*, location(1, 1, 0, 29, 2, 516)*/);
 
 module.exports = {
 	ast: a,