+++ test/cases/object-only.js
@@ -1,9 +1,9 @@
 var types = require('../types');
-var position = types.position;
+var location = types.location;
 var object = types.createObject;
 
 module.exports = {
-	ast: object([], position(1, 1, 0, 1, 3, 2)),
+	ast: object([], location(1, 1, 0, 1, 3, 2)),
 	options: {
 		verbose: true
 	}