+++ test/cases/string-only.js
@@ -1,9 +1,9 @@
 var types = require('../types');
-var position = types.position;
+var location = types.location;
 var string = types.createString;
 
 module.exports = {
-	ast: string('Some text', position(1, 1, 0, 1, 12, 11)),
+	ast: string('Some text', location(1, 1, 0, 1, 12, 11)),
 	options: {
 		verbose: true
 	}