+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## v2.0.0
+Added position.fileName
+Renamed position to loc
+Renamed position.char to loc.offset
+
 ## v1.2.15
 Fixed unicode parser bug